split connection lifecycle
