+++ client-g2/src/g2/runtime/g2-runtime.ts
@@ -11,25 +11,10 @@ import type { HttpClient } from "../../client/http";
 import { useAppStore } from "../../app/app-store";
 import { useSessionStore } from "../../client/session-store";
 import type { LogLine } from "../../client/session-format";
-import type { ConnectionStatus } from "../../client/ws";
-import type { ServerMessage } from "@float-code/shared/protocol";
-import { deriveUrls } from "../../constants";
 import type { G2Context } from "./g2-context";
 import type { G2State, RuntimeEvent } from "./g2-state";
-import { createConnectingState } from "../states/connecting/state";
 import { createErrorState } from "../states/error/state";
-import { createMainState } from "../states/main/state";
-import { createWorkspaceSelectState } from "../states/workspace-select/state";
-
-const LIFECYCLE_KEYS = {
-  wsError: "ws-error",
-  wsPairing: "ws-pairing",
-  wsDisconnected: "ws-disconnected",
-  authOk: "auth-ok",
-  authError: "auth-error",
-} as const;
-
-type LifecycleKey = (typeof LIFECYCLE_KEYS)[keyof typeof LIFECYCLE_KEYS];
+import { ConnectionLifecycle } from "./connection-lifecycle";
 
 function buildSttContext(messages: string[]): string {
   if (messages.length === 0) return "";
@@ -59,8 +44,7 @@ export class G2Runtime {
   private voiceSession: VoiceInputSession | null = null;
   private ctx: G2Context;
   private transitionChain = Promise.resolve();
-  private connectTimeoutTimer: ReturnType<typeof setTimeout> | null = null;
-  private pendingLifecycleKey: LifecycleKey | null = null;
+  private lifecycle: ConnectionLifecycle;
   private unsubscribeEvent: (() => void) | null = null;
   private unsubscribeWs: (() => void) | null = null;
   private unsubscribeWsStatus: (() => void) | null = null;
@@ -82,8 +66,15 @@ export class G2Runtime {
       startVoiceSession: (opts) => this.startVoiceSession(opts),
       stopVoiceSession: (reason) => this.stopVoiceSession(reason),
       getVoiceSession: () => this.voiceSession,
-      requestConnect: () => void this.requestConnect(),
+      requestConnect: () => void this.lifecycle.requestConnect(),
     };
+
+    this.lifecycle = new ConnectionLifecycle({
+      transition: (next) => this.transition(next),
+      getCurrentStateId: () => this.currentState?.id,
+      wsClient: options.wsClient,
+      httpClient: options.httpClient,
+    });
   }
 
   async start(): Promise<void> {
@@ -116,7 +107,7 @@ export class G2Runtime {
       this.dispatch({ kind: "ws", status });
     });
 
-    await this.requestConnect();
+    await this.lifecycle.requestConnect();
   }
 
   private async startVoiceSession(options?: {
@@ -160,7 +151,7 @@ export class G2Runtime {
   private dispatch(event: RuntimeEvent): void {
     Promise.resolve()
       .then(() => {
-        if (this.handleLifecycleEvent(event)) return;
+        if (this.lifecycle.intercept(event)) return;
         return this.currentState?.handle?.(this.ctx, event);
       })
       .catch((err) => {
@@ -264,7 +255,7 @@ export class G2Runtime {
   }
 
   dispose(): void {
-    this.clearConnectTimeout();
+    this.lifecycle.dispose();
     this.unsubscribeEvent?.();
     this.unsubscribeEvent = null;
     this.unsubscribeWs?.();
@@ -272,107 +263,5 @@ export class G2Runtime {
     this.unsubscribeWsStatus?.();
     this.unsubscribeWsStatus = null;
     this.ctx.wsClient.disconnect();
-  }
-
-  // --- Lifecycle management ---
-
-  private static readonly CONNECT_TIMEOUT_MS = 15_000;
-
-  private async requestConnect(): Promise<void> {
-    await this.transition(createConnectingState());
-
-    const { serverHost, serverToken } = useAppStore.getState();
-    if (!serverHost || !serverToken) {
-      await this.transition(
-        createErrorState("Please configure server in app settings"),
-      );
-      return;
-    }
-
-    const urls = deriveUrls(serverHost);
-    this.ctx.wsClient.updateConfig(urls.wsUrl, serverToken);
-    this.ctx.httpClient.updateConfig(urls.httpUrl, serverToken);
-
-    this.clearConnectTimeout();
-    this.connectTimeoutTimer = setTimeout(() => {
-      this.ctx.wsClient.disconnect();
-      void this.transition(createErrorState("Connection timeout"));
-    }, G2Runtime.CONNECT_TIMEOUT_MS);
-
-    this.ctx.wsClient.connect();
-  }
-
-  private clearConnectTimeout(): void {
-    if (this.connectTimeoutTimer) {
-      clearTimeout(this.connectTimeoutTimer);
-      this.connectTimeoutTimer = null;
-    }
-  }
-
-  private handleLifecycleEvent(event: RuntimeEvent): boolean {
-    if (event.kind === "ws") return this.handleLifecycleWs(event.status);
-    if (event.kind === "cc") return this.handleLifecycleCc(event.message);
-    return false;
-  }
-
-  private handleLifecycleWs(status: ConnectionStatus): boolean {
-    if (status.state === "error") {
-      this.clearConnectTimeout();
-      this.lifecycleTransition(LIFECYCLE_KEYS.wsError, () =>
-        createErrorState(status.reason),
-      );
-      return true;
-    }
-    if (status.state === "pairing") {
-      this.clearConnectTimeout();
-      this.lifecycleTransition(LIFECYCLE_KEYS.wsPairing, () =>
-        createErrorState(
-          `Pairing: ${status.code}\nApprove on server to connect`,
-        ),
-      );
-      return true;
-    }
-    if (
-      status.state === "disconnected" &&
-      this.currentState?.id !== "connecting"
-    ) {
-      this.lifecycleTransition(LIFECYCLE_KEYS.wsDisconnected, () =>
-        createErrorState("Disconnected"),
-      );
-      return true;
-    }
-    if (status.state === "connected") {
-      this.clearConnectTimeout();
-    }
-    return false;
-  }
-
-  private handleLifecycleCc(message: ServerMessage): boolean {
-    if (message.type === "auth.ok") {
-      this.lifecycleTransition(LIFECYCLE_KEYS.authOk, () =>
-        useSessionStore.getState().hasActive
-          ? createMainState()
-          : createWorkspaceSelectState(),
-      );
-      return true;
-    }
-    if (message.type === "auth.error") {
-      this.lifecycleTransition(LIFECYCLE_KEYS.authError, () =>
-        createErrorState(message.message),
-      );
-      return true;
-    }
-    return false;
-  }
-
-  private lifecycleTransition(
-    key: LifecycleKey,
-    createNext: () => G2State,
-  ): void {
-    if (this.pendingLifecycleKey === key) return;
-    this.pendingLifecycleKey = key;
-    void this.transition(createNext()).finally(() => {
-      if (this.pendingLifecycleKey === key) this.pendingLifecycleKey = null;
-    });
   }
 }
